refactor(title): unify fallback extractors behind TitleStrategy trait (#128 Debt #1)
Replace the ad-hoc if-let ladder of four free functions inside
`extract_title` with a small, explicit registry of strategies.

## Before

`title/mod.rs` (728 lines) contained four `extract_*` free functions
\u2014 `extract_cjk_bracket_title`, `extract_after_bracket_group`,
`extract_unclaimed_bracket_title`, `extract_title_from_parent` \u2014
each with a bespoke signature, called from `extract_title` via:

    if let Some(title) = extract_cjk_bracket_title(...)        { return Some(title); }
    if let Some(title) = extract_after_bracket_group(...)      { return Some(title); }
    if let Some(title) = extract_unclaimed_bracket_title(...)  { return Some(title); }
    return extract_title_from_parent(input, matches);

D10 (added in #129) names a 6th `extract_*` strategy as the
refactor-first tripwire. We were at 4; the next addition would have
hit it.

## After

New `title/strategies/` submodule with:

  trait TitleStrategy {
      fn name(&self) -> &'static str;
      fn try_extract(&self, ctx: &StrategyContext<'_>) -> Option<MatchSpan>;
  }

  static FALLBACK_STRATEGIES: &[&dyn TitleStrategy] = &[
      &CjkBracket, &AfterBracketGroup, &UnclaimedBracket, &ParentDir,
  ];

  fn run_fallback_ladder(ctx: &StrategyContext<'_>) -> Option<MatchSpan>;

Each strategy lives in its own file (77\u2013118 lines, avg ~99). Adding
a new strategy is now: implement the trait + one line in
`FALLBACK_STRATEGIES`. The orchestrator never changes.

## Behavior

Strictly neutral. Verified by:
- All 266 lib unit tests + every integration test pass unchanged
- `cargo clippy --all-targets -- -D warnings` clean
- `cargo fmt --check` clean
- `extract_title` body untouched apart from the if-let chain swap;
  every strategy's body is the previous free function's body verbatim.

## Diff highlights

- src/properties/title/mod.rs: 728 \u2192 426 (\u2212302 lines, well under
  the 600-line cohesion budget)
- src/properties/title/strategies/{after_bracket_group, cjk_bracket,
  parent_dir, unclaimed_bracket}.rs: 4 new files, 77\u2013118 lines each
- src/properties/title/strategies/mod.rs: trait + registry + ladder
  runner with module-level docs explaining ordering and the
  add-a-strategy procedure

Net diff is +192 lines for the project, but every added line is
either module docs or reusable trait/registry scaffolding. The
production logic per strategy is identical.

## Bonus added value

`run_fallback_ladder` emits a `log::trace!` for every strategy hit
naming the strategy that claimed the title \u2014 previously this was
opaque (just "some fallback returned a title").

Refs #128. No public API changes.
